+++ source/src/vertmodel.h
@@ -350,7 +350,7 @@ __attribute__((optimize(2)))
                 }
             }
 
-            if(dbgstenc >= (owner->numframes > 1 || as.anim&ANIM_DYNALLOC ? 2 : 1)) conoutf("%s: %d tris", owner->filename, (idx - d->idxs())/3);
+            if(dbgstenc >= (owner->numframes > 1 || as.anim&ANIM_DYNALLOC ? 2 : 1)) conoutf("%s: %d tris", owner->filename, int((idx - d->idxs()) / 3));
 
             d->size = (uchar *)idx - (uchar *)d;
             return d;
